requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cloudscraper==1.2.60
garminconnect==0.1.44
google-api-core==2.7.1
google-auth==2.6.2
google-cloud-secret-manager==2.9.2
googleapis-common-protos==1.56.0
grpc-google-iam-v1==0.12.3
grpcio==1.45.0
grpcio-status==1.45.0
idna==3.3
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
